+++ examples/nineslice.rs
@@ -1,5 +1,6 @@
 use tetra::graphics::ui::NineSlice;
-use tetra::graphics::{self, Color, Rectangle, Texture, Vec2};
+use tetra::graphics::{self, Color, Rectangle, Texture};
+use tetra::math::Vec2;
 use tetra::{Context, Game, State};
 
 struct GameState {