+++ examples/text_input.rs
@@ -1,5 +1,6 @@
-use tetra::graphics::{self, Color, Font, Text, Vec2};
+use tetra::graphics::{self, Color, Font, Text};
 use tetra::input::{self, Key};
+use tetra::math::Vec2;
 use tetra::{Context, Game, State};
 
 struct GameState {