+++ examples/texture.rs
@@ -1,4 +1,5 @@
-use tetra::graphics::{self, Color, DrawParams, Texture, Vec2};
+use tetra::graphics::{self, Color, DrawParams, Texture};
+use tetra::math::Vec2;
 use tetra::{Context, Game, State};
 
 struct GameState {