+++ src/glm.rs
@@ -1,9 +1,0 @@
-//! Functions and types relating to math (provided by the nalgebra-glm crate).
-//!
-//! Ideally, the documentation for `nalgebra-glm` would be reproduced here for your convienence. Unfortunately,
-//! [rustdoc currently doesn't handle this very well](https://github.com/rust-lang/rust/issues/15823), and the
-//! re-exported docs would likely be more confusing than helpful. Until those issues are fixed, you can find
-//! the documentation for `nalgebra-glm` by clicking the re-export link below.
-
-#[doc(no_inline)]
-pub use nalgebra_glm::*;